Rename canvas to backdrop, rearrange and clean up code in Backdrop.js, clean up imports
